feat(main, levellist): prototypes of creating contact, viewing contact appoinments and creating appointment
(main.c): separated some code into functions
(levellist.h, levellist.c): created a function getElementPtrLevelList which for now is a copy of
getElementIndexLevelList except it returns a pointer to the cell instead
of the index of the cell. This function currently has O(n) complexity which is
bad, it should use dichotomic search instead.